+++ storage/Cargo.toml
@@ -102,15 +102,15 @@ path = "src/qmdb/benches/bench.rs"
 required-features = ["test-traits"]
 
 [[bench]]
-name = "init_scale"
+name = "qmdb_gungraun"
 harness = false
-path = "src/qmdb/benches/init_scale.rs"
+path = "src/qmdb/benches/bench_gungraun.rs"
 required-features = ["test-traits"]
 
 [[bench]]
-name = "qmdb_gungraun"
+name = "init_scale"
 harness = false
-path = "src/qmdb/benches/gungraun.rs"
+path = "src/qmdb/benches/init_scale.rs"
 required-features = ["test-traits"]
 
 [[bench]]
